fix: Fix truth matching in `TruthVertexFinder` in Examples (#2937)
Our vertex truth finding relied on the fact that we have one track for each selected particle which can only be achieved in some truth tracking cases.

I generalized this by truth matching the tracks first and then grouping the tracks to the truth vertices.

This can be used to isolate vertex fitter performance.

Note that the fitter will always be full billoir for now. This could be extended to AMVF in the future.